requirements.txt
awscli
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws-kms
aws-cdk.aws-iam
aws-cdk.aws-secretsmanager
aws-cdk.aws-rds
aws-cdk.aws-elasticache
